remove unchanged imports from go-git
