+++ pkg/ignore/dir.go
@@ -1,157 +1,85 @@
 package ignore
 
 import (
-	"bufio"
-	"bytes"
-	"errors"
-	"io/ioutil"
+	"fmt"
 	"os"
+	"path/filepath"
 	"strings"
 
-	"github.com/go-git/go-billy/v5"
-	"github.com/go-git/go-git/v5/plumbing/format/config"
-	gioutil "github.com/go-git/go-git/v5/utils/ioutil"
-	"github.com/rs/zerolog/log"
+	"github.com/get-woke/woke/pkg/util"
 )
 
-const (
-	commentPrefix = "#"
-	coreSection   = "core"
-	excludesfile  = "excludesfile"
-	gitDir        = ".git"
-	gitconfigFile = ".gitconfig"
-	systemFile    = "/etc/gitconfig"
-)
-
-var defaultIgnoreFiles = []string{
-	".gitignore",
-	".wokeignore",
-}
-
-// readIgnoreFile reads a specific git ignore file.
-func readIgnoreFile(fs billy.Filesystem, path []string, ignoreFile string) (ps []Pattern, err error) {
-	f, err := fs.Open(fs.Join(append(path, ignoreFile)...))
+func getRootFileSystem(findRootDir bool) (string, []string, error) {
+	var domain []string
+	cwd, err := os.Getwd()
 	if err != nil {
-		_event := log.Warn()
-		if errors.Is(err, os.ErrNotExist) {
-			_event = log.Debug()
-			err = nil
-		}
-		_event.Err(err).Str("file", fs.Join(append(path, ignoreFile)...)).Msg("skipping ignorefile")
-		return []Pattern{}, err
+		fmt.Println(err)
+		return "", domain, err
 	}
-
-	defer f.Close()
-
-	scanner := bufio.NewScanner(f)
-	for scanner.Scan() {
-		s := scanner.Text()
-		if !strings.HasPrefix(s, commentPrefix) && len(strings.TrimSpace(s)) > 0 {
-			ps = append(ps, ParsePattern(s, path))
-		}
-	}
-
-	return
-}
-
-// ReadPatterns reads gitignore patterns recursively traversing through the directory
-// structure. The result is in the ascending order of priority (last higher).
-func ReadPatterns(fs billy.Filesystem, path []string) (ps []Pattern, err error) {
-	ps = []Pattern{}
-	for _, filename := range defaultIgnoreFiles {
-		var subps []Pattern
-		subps, err = readIgnoreFile(fs, path, filename)
+	if findRootDir {
+		rootDir, err := detectGitPath(".")
 		if err != nil {
-			return ps, err
+			fmt.Println(err)
+			return "", domain, err
 		}
-		if len(subps) > 0 {
-			ps = append(ps, subps...)
-		}
+		domain = strings.Split(util.ParseRelativePathFromAbsolutePath(cwd, rootDir), string(os.PathSeparator))
+		return rootDir, domain, nil
 	}
+	return cwd, domain, nil
+}
 
-	var fis []os.FileInfo
-	fis, err = fs.ReadDir(fs.Join(path...))
+func detectGitPath(path string) (string, error) {
+	// normalize the path
+	path, err := filepath.Abs(path)
 	if err != nil {
-		return ps, err
+		return "", err
 	}
 
-	for _, fi := range fis {
-		if fi.IsDir() && fi.Name() != gitDir {
-			var subps []Pattern
-			subps, err = ReadPatterns(fs, append(path, fi.Name()))
-			if err != nil {
-				return ps, err
+	for {
+		fi, err := os.Stat(filepath.Join(path, gitDir))
+		if err == nil {
+			if !fi.IsDir() {
+				return "", fmt.Errorf("%v exist but is not a directory", gitDir)
 			}
+			return path, nil
+		}
+		if !os.IsNotExist(err) {
+			// unknown error
+			return "", err
+		}
 
-			if len(subps) > 0 {
-				ps = append(ps, subps...)
-			}
+		// detect bare repo
+		ok, err := isGitDir(path)
+		if err != nil {
+			return "", err
+		}
+		if ok {
+			return path, nil
+		}
+
+		if parent := filepath.Dir(path); parent == path {
+			return "", fmt.Errorf(".git not found")
+		} else {
+			path = parent
+		}
+	}
+}
+
+func isGitDir(path string) (bool, error) {
+	markers := []string{"HEAD", "objects", "refs"}
+
+	for _, marker := range markers {
+		_, err := os.Stat(filepath.Join(path, marker))
+		if err == nil {
+			continue
+		}
+		if !os.IsNotExist(err) {
+			// unknown error
+			return false, err
+		} else {
+			return false, nil
 		}
 	}
 
-	return ps, nil
-}
-
-func loadPatterns(fs billy.Filesystem, path string) (ps []Pattern, err error) {
-	f, err := fs.Open(path)
-	if err != nil {
-		if os.IsNotExist(err) {
-			return nil, nil
-		}
-		return nil, err
-	}
-
-	defer gioutil.CheckClose(f, &err)
-
-	b, err := ioutil.ReadAll(f)
-	if err != nil {
-		return ps, err
-	}
-
-	d := config.NewDecoder(bytes.NewBuffer(b))
-
-	raw := config.New()
-	if err = d.Decode(raw); err != nil {
-		return ps, err
-	}
-
-	s := raw.Section(coreSection)
-	efo := s.Options.Get(excludesfile)
-	if efo == "" {
-		return nil, nil
-	}
-
-	ps, err = readIgnoreFile(fs, nil, efo)
-	if os.IsNotExist(err) {
-		return nil, nil
-	}
-
-	return ps, nil
-}
-
-// LoadGlobalPatterns loads gitignore patterns from from the gitignore file
-// declared in a user's ~/.gitconfig file.  If the ~/.gitconfig file does not
-// exist the function will return nil.  If the core.excludesfile property
-// is not declared, the function will return nil.  If the file pointed to by
-// the core.excludesfile property does not exist, the function will return nil.
-//
-// The function assumes fs is rooted at the root filesystem.
-func LoadGlobalPatterns(fs billy.Filesystem) (ps []Pattern, err error) {
-	home, err := os.UserHomeDir()
-	if err != nil {
-		return
-	}
-
-	return loadPatterns(fs, fs.Join(home, gitconfigFile))
-}
-
-// LoadSystemPatterns loads gitignore patterns from from the gitignore file
-// declared in a system's /etc/gitconfig file.  If the /etc/gitconfig file does
-// not exist the function will return nil.  If the core.excludesfile property
-// is not declared, the function will return nil.  If the file pointed to by
-// the core.excludesfile property does not exist, the function will return nil.
-//
-// The function assumes fs is rooted at the root filesystem.
-func LoadSystemPatterns(fs billy.Filesystem) (ps []Pattern, err error) {
-	return loadPatterns(fs, systemFile)
+	return true, nil
 }
