+++ pkg/ignore/ignore.go
@@ -1,54 +1,98 @@
 package ignore
 
 import (
+	"bufio"
+	"errors"
 	"fmt"
 	"os"
-	"path/filepath"
 	"strings"
 	"time"
 
+	"github.com/go-git/go-billy/v5"
 	"github.com/go-git/go-billy/v5/osfs"
+	"github.com/go-git/go-git/v5/plumbing/format/gitignore"
 	"github.com/rs/zerolog/log"
+
+	"github.com/get-woke/woke/pkg/util"
 )
 
 // Ignore is a gitignore-style object to ignore files/directories
 type Ignore struct {
-	matcher Matcher
+	matcher gitignore.Matcher
 }
 
-// given a absolute path (example: /runner/folder/root/data/effx.yaml)
-// and given a workingDir (example: root)
-// it will return data/effx.yaml
-func parseRelativePathFromAbsolutePath(absoluteDir, workingDir string) string {
-	// if working directory does not end with a slash, add it
-	if strings.LastIndex(workingDir, "/") != len(workingDir)-1 {
-		workingDir += "/"
-	}
+const (
+	commentPrefix = "#"
+	gitDir        = ".git"
+)
 
-	res := strings.Split(absoluteDir, workingDir)
-	if len(res) > 1 {
-		return res[1]
-	}
-	return ""
+var defaultIgnoreFiles = []string{
+	".gitignore",
+	".wokeignore",
 }
 
-func getRootFileSystem(findRootDir bool) (string, []string, error) {
-	var domain []string
-	cwd, err := os.Getwd()
+// readIgnoreFile reads a specific git ignore file.
+func readIgnoreFile(fs billy.Filesystem, path []string, ignoreFile string) (ps []gitignore.Pattern, err error) {
+	f, err := fs.Open(fs.Join(append(path, ignoreFile)...))
 	if err != nil {
-		fmt.Println(err)
-		return "", domain, err
-	}
-	if findRootDir {
-		rootDir, err := detectGitPath(".")
-		if err != nil {
-			fmt.Println(err)
-			return "", domain, err
+		_event := log.Warn()
+		if errors.Is(err, os.ErrNotExist) {
+			_event = log.Debug()
+			err = nil
 		}
-		domain = strings.Split(parseRelativePathFromAbsolutePath(cwd, rootDir), string(os.PathSeparator))
-		return rootDir, domain, nil
+		_event.Err(err).Str("file", fs.Join(append(path, ignoreFile)...)).Msg("skipping ignorefile")
+		return []gitignore.Pattern{}, err
 	}
-	return cwd, domain, nil
+
+	defer f.Close()
+
+	scanner := bufio.NewScanner(f)
+	for scanner.Scan() {
+		s := scanner.Text()
+		if !strings.HasPrefix(s, commentPrefix) && len(strings.TrimSpace(s)) > 0 {
+			ps = append(ps, gitignore.ParsePattern(s, path))
+		}
+	}
+
+	return
+}
+
+// readPatterns reads gitignore patterns recursively traversing through the directory
+// structure. The result is in the ascending order of priority (last higher).
+func readPatterns(fs billy.Filesystem, path []string) (ps []gitignore.Pattern, err error) {
+	ps = []gitignore.Pattern{}
+	for _, filename := range defaultIgnoreFiles {
+		var subps []gitignore.Pattern
+		subps, err = readIgnoreFile(fs, path, filename)
+		if err != nil {
+			return ps, err
+		}
+		if len(subps) > 0 {
+			ps = append(ps, subps...)
+		}
+	}
+
+	var fis []os.FileInfo
+	fis, err = fs.ReadDir(fs.Join(path...))
+	if err != nil {
+		return ps, err
+	}
+
+	for _, fi := range fis {
+		if fi.IsDir() && fi.Name() != gitDir {
+			var subps []gitignore.Pattern
+			subps, err = readPatterns(fs, append(path, fi.Name()))
+			if err != nil {
+				return ps, err
+			}
+
+			if len(subps) > 0 {
+				ps = append(ps, subps...)
+			}
+		}
+	}
+
+	return ps, nil
 }
 
 // NewIgnore produces an Ignore object, with compiled lines from defaultIgnoreFiles
@@ -68,92 +112,26 @@ func NewIgnore(lines []string, findRootDir bool) *Ignore {
 	}
 	rootFs := osfs.New(rootDir)
 	currentPath := []string{"."}
-	ps, err := ReadPatterns(rootFs, currentPath)
+	ps, err := readPatterns(rootFs, currentPath)
 	if err != nil {
 		fmt.Println(err)
 		return nil
 	}
 
 	for _, line := range lines {
-		pattern := ParsePattern(line, domain)
+		pattern := gitignore.ParsePattern(line, domain)
 		ps = append(ps, pattern)
 	}
 
 	ignorer := Ignore{
-		matcher: NewMatcher(ps),
+		matcher: gitignore.NewMatcher(ps),
 	}
 
 	return &ignorer
 }
 
-func filterEmptyStrings(s []string) []string {
-	var r []string
-	for _, str := range s {
-		if str != "" {
-			r = append(r, str)
-		}
-	}
-	return r
-}
-
 // Match returns true if the provided file matches any of the defined ignores
 func (i *Ignore) Match(f string, isDir bool) bool {
-	parts := filterEmptyStrings(strings.Split(f, string(os.PathSeparator)))
+	parts := util.FilterEmptyStrings(strings.Split(f, string(os.PathSeparator)))
 	return i.matcher.Match(parts, isDir)
-}
-
-func detectGitPath(path string) (string, error) {
-	// normalize the path
-	path, err := filepath.Abs(path)
-	if err != nil {
-		return "", err
-	}
-
-	for {
-		fi, err := os.Stat(filepath.Join(path, gitDir))
-		if err == nil {
-			if !fi.IsDir() {
-				return "", fmt.Errorf("%v exist but is not a directory", gitDir)
-			}
-			return path, nil
-		}
-		if !os.IsNotExist(err) {
-			// unknown error
-			return "", err
-		}
-
-		// detect bare repo
-		ok, err := isGitDir(path)
-		if err != nil {
-			return "", err
-		}
-		if ok {
-			return path, nil
-		}
-
-		if parent := filepath.Dir(path); parent == path {
-			return "", fmt.Errorf(".git not found")
-		} else {
-			path = parent
-		}
-	}
-}
-
-func isGitDir(path string) (bool, error) {
-	markers := []string{"HEAD", "objects", "refs"}
-
-	for _, marker := range markers {
-		_, err := os.Stat(filepath.Join(path, marker))
-		if err == nil {
-			continue
-		}
-		if !os.IsNotExist(err) {
-			// unknown error
-			return false, err
-		} else {
-			return false, nil
-		}
-	}
-
-	return true, nil
 }
