+++ pkg/util/string.go
@@ -1,6 +1,9 @@
 package util
 
-import "fmt"
+import (
+	"fmt"
+	"strings"
+)
 
 // MarkdownCodify returns a markdown code string
 // https://www.markdownguide.org/basic-syntax/#code
@@ -19,4 +22,30 @@ func InSlice(s string, slice []string) bool {
 		}
 	}
 	return false
+}
+
+func FilterEmptyStrings(s []string) []string {
+	var r []string
+	for _, str := range s {
+		if str != "" {
+			r = append(r, str)
+		}
+	}
+	return r
+}
+
+// given a absolute path (example: /runner/folder/root/data/effx.yaml)
+// and given a workingDir (example: root)
+// it will return data/effx.yaml
+func ParseRelativePathFromAbsolutePath(absoluteDir, workingDir string) string {
+	// if working directory does not end with a slash, add it
+	if strings.LastIndex(workingDir, "/") != len(workingDir)-1 {
+		workingDir += "/"
+	}
+
+	res := strings.Split(absoluteDir, workingDir)
+	if len(res) > 1 {
+		return res[1]
+	}
+	return ""
 }
